get data update

+++ flutter_hotel_booking_ui/packages/hotel_repository/lib/src/entities/hotel_entity.dart
@@ -14,8 +14,8 @@ class HotelEntity {
   Timestamp date;
   // DateText dateTxt;
   // RoomData roomData;
-  int dist;
-  int rating;
+  double dist;
+  double rating;
   int reviews;
   int perNight;
   bool isSelected;
