+++ flutter_hotel_booking_ui/packages/hotel_repository/lib/src/models/hotel.dart
@@ -13,8 +13,8 @@ class Hotel {
   Timestamp date;
   //DateText dateTxt;
   //RoomData roomData;
-  int dist;
-  int rating;
+  double dist;
+  double rating;
   int reviews;
   int perNight;
   bool isSelected;
